Fetching initital userId in layout so we can remove useEffect in UserIdContext
