+++ src/lib/dataService.ts
@@ -1,8 +1,7 @@
 import { supabase } from './supabaseClient';
 import { Todo } from '../../types';
-import { authOptions } from "../lib/authOptions";
-import { getServerSession } from "next-auth";
 import { renderSanitizedMarkdown } from "./markdown";
+import { getAuthenticatedUserId } from './userService';
 
 type ReorderUpdateInput = {
   id: string;
@@ -47,29 +46,6 @@ function normalizeComparableId(value: string | number | null | undefined): strin
   if (value === null || typeof value === 'undefined') return null;
   const normalized = String(value);
   return normalized.length > 0 ? normalized : null;
-}
-
-export async function fetchUserIdByEmail(email: string): Promise<number> {
-  const baseUrl = process.env.NEXT_PUBLIC_BASE_URL;
-  if (!baseUrl) {
-    throw new Error("NEXT_PUBLIC_BASE_URL environment variable is not set");
-  }
-  const userIdRes = await fetch(
-    `${baseUrl}/api/userid?email=${encodeURIComponent(email)}`,
-    { cache: "no-store" }
-  );
-  if (!userIdRes.ok) throw new Error("Could not fetch user id");
-  const { userId } = await userIdRes.json();
-  if (typeof userId !== 'number') throw new Error('userId must be a number');
-  return userId;
-}
-
-async function getAuthenticatedUserId(): Promise<number> {
-  const session = await getServerSession(authOptions);
-  if (!session) throw new Error("User not authenticated");
-  const email = session.user?.email;
-  if (!email) throw new Error("User email missing");
-  return fetchUserIdByEmail(email);
 }
 
 function normalizeSortIndex(value: number | null | undefined): number {
